# cmd/api/docs/swagger.yaml
definitions:
  models.AddCoupon:
    properties:
      coupon:
        type: string
      discount_percentage:
        type: integer
      minimum_price:
        type: number
      validity:
        type: boolean
    required:
    - coupon
    - discount_percentage
    - minimum_price
    - validity
    type: object
...
  models.CategoryOfferReceiver:
    properties:
      category_id:
        type: integer
      discount_percentage:
        type: integer
      offer_name:
        type: string
    required:
    - category_id
    - discount_percentage
    - offer_name
    type: object
...
  models.CouponAddUser:
    properties:
      coupon_name:
        type: string
    required:
    - coupon_name
    type: object
...
  models.ProductOfferReceiver:
    properties:
      discount_percentage:
        type: integer
      offer_name:
        type: string
      product_id:
        type: integer
    required:
    - discount_percentage
    - offer_name
    - product_id
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Zhooze is an E-commerce platform to purchase and sell shoes
  title: Go + Gin Zhooze E-Commerce API
  version: 1.0.0
paths:
...
  /admin/coupons:
    get:
      consumes:
      - application/json
      description: Get Available coupon details for admin side
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get coupon details
      tags:
      - Admin Coupon Management
    patch:
      consumes:
      - application/json
      description: Expire Coupon by admin which are already present by passing coupon
        id
      parameters:
      - description: Coupon id
        in: query
        name: coupon_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Expire Coupon
      tags:
      - Admin Coupon Management
    post:
      consumes:
      - application/json
      description: Add A new Coupon which can be used by the users from the checkout
        section
      parameters:
      - description: Add new Coupon
        in: body
        name: coupon
        required: true
        schema:
          $ref: '#/definitions/models.AddCoupon'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add  a new coupon by Admin
      tags:
      - Admin Coupon Management
...
  /admin/offer/category-offer:
    delete:
      consumes:
      - application/json
      description: Add a new Offer for a category by specifying a limit
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add  Category Offer
      tags:
      - Admin Offer Management
    get:
      consumes:
      - application/json
      description: Add a new Offer for a category by specifying a limit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add  Category Offer
      tags:
      - Admin Offer Management
    post:
      consumes:
      - application/json
      description: Add a new Offer for a product by specifying a limit
      parameters:
      - description: Add new category Offer
        in: body
        name: coupon
        required: true
        schema:
          $ref: '#/definitions/models.CategoryOfferReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Category Offer
      tags:
      - Admin Offer Management
  /admin/offer/product-offer:
    delete:
      consumes:
      - application/json
      description: Add a new Offer for a product by specifying a limit
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add  Product Offer
      tags:
      - Admin Offer Management
    get:
      consumes:
      - application/json
      description: Add a new Offer for a product by specifying a limit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add  Product Offer
      tags:
      - Admin Offer Management
    post:
      consumes:
      - application/json
      description: Add a new Offer for a product by specifying a limit
      parameters:
      - description: Add new Product Offer
        in: body
        name: coupon
        required: true
        schema:
          $ref: '#/definitions/models.ProductOfferReceiver'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add  Product Offer
      tags:
      - Admin Offer Management
...
  /user/coupon/apply:
    post:
      consumes:
      - application/json
      description: Add coupon to get discount on Checkout section
      parameters:
      - description: Add coupon to order
        in: body
        name: couponDetails
        required: true
        schema:
          $ref: '#/definitions/models.CouponAddUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Apply coupon on Checkout Section
      tags:
      - User Checkout
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"